refactored cop cards into the landing page
