--- app/frontend/src/components/Cards/_InnerCopCard.scss
@@ -1,20 +1,0 @@
-// Internal Imports
-@use "../Basics" as *;
-
-.inner-cop-card {
-  background-color: $color-white;
-  height: 100%;
-  max-width: 625px;
-  min-height: 448px;
-}
-
-.inner-cop-card-content {
-  box-sizing: border-box;
-  height: 100%;
-  padding: {
-    top: 8px;
-    right: 0.6vw;
-    left: 0.6vw;
-    bottom: 8px;
-  }
-}
