--- app/frontend/src/components/Cards/_InnerCopNavCard.scss
@@ -1,19 +1,0 @@
-// Internal Imports
-@use "../Basics" as *;
-
-.inner-cop-nav-card {
-  background-color: $color-white;
-  border: none;
-  border-radius: 16px;
-  max-width: 188px;
-  min-height: 64px;
-
-  &.active {
-    background-color: $color-blue-dark;
-    color: $color-white;
-  }
-
-  &:hover {
-    cursor: pointer;
-  }
-}
