+++ app/frontend/src/index.scss
@@ -7,9 +7,6 @@
 // Import component styles
 @use "components/Buttons" as *;
 @use "components/Cards/StandardCard";
-@use "components/Cards/CopCard";
-@use "components/Cards/InnerCopCard";
-@use "components/Cards/InnerCopNavCard";
 @use "components/Cards/CircleCard";
 @use "components/Carousel/ScrollCarousel";
 @use "components/Dialog/Dialog";
