+++ app/frontend/src/pages/LandingPage/_LandingPageCopCards.scss
@@ -1,5 +1,5 @@
 // Internal Imports
-@use "../Basics" as *;
+@use "components/Basics" as *;
 
 /**********************************
 *** Classes for large COP cards ***
@@ -37,7 +37,6 @@
 **********************************/
 .cop-card-sm {
   background-color: $color-grey-light;
-  box-shadow: 0 8px 8px rgb(0 0 0 / 25%);
   margin: 8px;
   max-width: 312px;
   min-height: 600px;
@@ -60,5 +59,48 @@
     top: 74px;
     right: 6.4vw;
     left: 6.4vw;
+  }
+}
+
+/**********************************
+*** Classes for Inner COP cards ***
+**********************************/
+
+.inner-cop-card {
+  background-color: $color-white;
+  height: 100%;
+  max-width: 625px;
+  min-height: 448px;
+}
+
+.inner-cop-card-content {
+  box-sizing: border-box;
+  height: 100%;
+  padding: {
+    top: 8px;
+    right: 0.6vw;
+    left: 0.6vw;
+    bottom: 8px;
+  }
+}
+
+/**************************************
+*** Classes for Inner COP Nav cards ***
+**************************************/
+
+.inner-cop-nav-card {
+  background-color: $color-white;
+  border: none;
+  border-radius: 16px;
+  max-width: 188px;
+  min-height: 64px;
+
+  &.active {
+    background-color: $color-blue-dark;
+    color: $color-white;
+  }
+
+  &:hover {
+    cursor: pointer;
   }
 }
